feat(zksync-os-sequencer): extract shareable components to `lib.rs`
